+++ integration_tests/files/problematic.py
@@ -1,0 +1,16 @@
+
+def problematic_function(arg1):
+    """Returns double the value of the first argument.
+
+    Args:
+        a: The first argument.
+
+    Returns:
+        Two times the value of the first argument.
+
+    """
+    return arg1 * 2
+
+
+def problematic_rst():
+    """"""